fixed issue in the nobit flip test for the Hopfield implementation
DeltaE is now of type int in all model

+++ codes/include/hopfield_nobits.hpp
@@ -27,7 +27,7 @@ public:
 
 private:
 
-    double DeltaE(int spin, int realization);
+    int DeltaE(int spin, int realization);
 
     void runSweepsSharedRNG(gsl_rng* ran, bool save, double freq);
     void runSweepsIndependentRNG(gsl_rng* ran, bool save, double freq);
